requirements.txt
hydra-core
mat73
matplotlib
numpy
omegaconf
Pillow
pytorch_lightning
scikit_image
scipy
torch
torchvision
tqdm
wandb